feat([slug]): add [slug].astro as SSR blogpost page, add 404 page and client side error handling to [slug].astro, fix nav-mobilemenu's class in navbar
