+++ src/data/blogposts/post-2.md
@@ -7,13 +7,13 @@ blogpostImg: /my-mini-blog/blogpost_img/default.jpg
 blogthumbImg: /my-mini-blog/blogthumb_img/default.jpg
 bannerImg: /my-mini-blog/banner_img/default.jpg
 content: Insert long sentences here, 26 characters at minimum.
-tags: ['computer science', 'journaling']
+tags: ['computer science', 'real life experience', 'journaling']
 createdAt: 2024-02-24T12:08:46.028Z
 updatedAt: 2024-02-24T12:08:46.028Z
 user: {
     _id: '655c7a2bd2ffdd11ea492d1c',
     name: 'Barry T. Quokka',
-    photo: "/my-mini-blog/user/default.jpg"
+    photo: "/my-mini-blog/user/profile-655c7a2bd2ffdd11ea492d1c.jpg"
 }
 commentCount: 0
 ---
