+++ src/layouts/MainLayout.astro
@@ -15,7 +15,7 @@ const {title = "My Mini Blog"} = Astro.props as Props;
 
 <html lang="en">
 	<MainHead {title}/>
-	<body class="bg-neutral-00 relative w-full min-h-dvh overflow-x-hidden">
+	<body class="bg-neutral-00 relative w-full min-h-dvh overflow-x-hidden flex flex-col">
 		<NavigationBar />
 		<slot/>
 		<script src="../scripts/setNavbar.js"></script>
