+++ src/pages/404.astro
@@ -1,0 +1,22 @@
+---
+import MainHead from "@layouts/MainHead.astro"
+import "@styles/global.css";
+import "@styles/auth.css";
+---
+
+<html lang="en">
+    <MainHead title="Page is not found!"/>
+    <body class="bg-theme-blue sm:bg-neutral-00 w-full h-screen flex justify-center items-center sm:px-5">
+        <main class="auth-container">
+            <h2 class="auth-title">
+                Page is not found!
+            </h2>
+            <p class="text-white text-lg mb-3">
+                You may have mistyped the address or the page may have moved.
+            </p>
+            <div class="w-full flex justify-center text-center">
+                <a href="/" class="auth-button auth-loading">go back to home&nbsp;page</a>
+            </div>
+        </main>
+    </body>
+</html>
