+++ src/styles/auth.css
@@ -2,7 +2,7 @@
     height: fit-content;
     width: 100%;
     max-width: 25rem;
-    margin-bottom: 6rem;
+    margin-bottom: 3.5rem;
     padding: 1.75rem;
 }
